chore: move tests to one separate "delegated.go" file
