+++ test/e2e_env/kubernetes/kubernetes_suite_test.go
@@ -56,6 +56,7 @@ var (
 	_ = Describe("Gateway - Gateway API", gateway.GatewayAPI, Ordered)
 	_ = Describe("Gateway - mTLS", gateway.Mtls, Ordered)
 	_ = Describe("Gateway - Resources", gateway.Resources, Ordered)
+	_ = Describe("Delegated Gateway", Label("kind-not-supported", "ipv6-not-supported"), gateway.Delegated, Ordered)
 	_ = Describe("Graceful", graceful.Graceful, Ordered)
 	_ = Describe("Eviction", graceful.Eviction, Ordered)
 	_ = XDescribe("Change Service", graceful.ChangeService, Ordered)
@@ -81,7 +82,6 @@ var (
 	_ = Describe("MeshHealthCheck API", meshhealthcheck.API, Ordered)
 	_ = Describe("MeshCircuitBreaker API", meshcircuitbreaker.API, Ordered)
 	_ = Describe("MeshCircuitBreaker", meshcircuitbreaker.MeshCircuitBreaker, Ordered)
-	_ = Describe("MeshCircuitBreaker Delegated Gateway", meshcircuitbreaker.MeshCircuitBreakerDelegatedGateway, Ordered)
 	_ = Describe("MeshMetric", meshmetric.MeshMetric, Ordered)
 	_ = Describe("MeshRetry", meshretry.API, Ordered)
 	_ = Describe("MeshProxyPatch", meshproxypatch.MeshProxyPatch, Ordered)
